Inference.Conjugate

Documentation

class Distribution a where

Describes a family of distributions with a fixed form. For example, a Bernoulli distribution is parameterized by a probability p and produces binary samples (True with probability p, False with probability 1-p).

Its Distribution instance is: > instance Distribution Bernoulli where > type Params Bernoulli = Double > type Support Bernoulli = Bool > distSample _ = uncurry bernoulli > distLogP _ p True = log p > distLogP _ p False = log (1 - p)

Associated Types

type Params a :: Type

type Support a :: Type

Methods

distSample :: PrimMonad m => a -> Params a -> Prob m (Support a)

distLogP :: a -> Params a -> Support a -> Double

newtype AsPrior p

A type-level marker for treating a distribution as a prior.

Constructors

AsPrior p

class (Distribution p, Distribution l, Support p ~ Params l) => Conjugate p l where

Marks two distributions as a conjugate pair of prior and likelihood. The property of such a pair is that the posterior has the same form as the prior (including the same Params and Support), and that its parameters can be obtained analytically from the parameters of the prior and a set of observations.

The class method updatePrior returns the parameters of the posterior given the prior parameters after a single observation.

Methods

priorSingleton

Arguments

:: p
provides a singleton instance of the prior distribution in order to make sampling from priors easier.

updatePrior :: l -> Params p -> Support l -> Params p

predLogP :: l -> Params p -> Support l -> Double

type family Hyper (a :: k) :: Type

type family Probs (a :: k) :: Type

newtype HyperRep p

Constructors

HyperRep

Fields

runHyper :: Hyper (AsPrior p)

newtype ProbsRep p

Constructors

ProbsRep

Fields

runProbs :: Probs (AsPrior p)

class Jeffreys a where

Methods

jeffreysPrior :: Hyper a

class GJeffreys t where

Methods

gjeffreysPrior :: forall p. t p

class Uniform a where

Methods

uniformPrior :: Hyper a

class GUniform t where

Methods

guniformPrior :: forall p. t p

class Prior a where

Methods

sampleProbs :: PrimMonad m => Hyper a -> Prob m (Probs a)

class GPrior i o where

Methods

gsampleProbs :: forall m p. PrimMonad m => i p -> Prob m (o p)

type Accessor r p = forall f. Lens' (r f) (f p)

class Monad m => RandomInterpreter m r | m -> r where

Associated Types

type SampleCtx m a :: Constraint

Methods

sampleValue :: (Conjugate p l, SampleCtx m l) => String -> l -> Accessor r p -> m (Support l)

sampleConst :: (Distribution d, SampleCtx m d) => String -> d -> Params d -> m (Support d)

permutationPlate :: Ord a => Int -> m a -> m [a]

newtype Trace (r :: (Type -> Type) -> Type)

Constructors

Trace

Fields

runTrace :: Seq Dynamic

observeValue :: (Conjugate p l, Typeable (Support l), Monad m) => String -> l -> Accessor r p -> Support l -> StateT (Trace r) m ()

observeConst :: (Distribution d, Typeable (Support d), Monad m) => String -> d -> Params d -> Support d -> StateT (Trace r) m ()

takeTrace :: Typeable a => Trace r -> Maybe (a, Trace r)

newtype SampleI m r a

Constructors

SampleI (ReaderT (r ProbsRep) (Prob m) a)

sampleResult :: p ProbsRep -> SampleI m p a -> Gen (PrimState m) -> m a

newtype TraceI m r a

Constructors

TraceI (ReaderT (r ProbsRep) (StateT (Trace r) (Prob m)) a)

sampleTrace :: r ProbsRep -> TraceI m r a -> Gen (PrimState m) -> m (a, Trace r)

newtype EvalTraceI r a

Constructors

EvalTraceI (ReaderT (r ProbsRep) (StateT (Trace r, Double) Maybe) a)

runTraceLogP :: r ProbsRep -> Trace r -> EvalTraceI r a -> Maybe (a, (Trace r, Double))

evalTraceLogP :: r ProbsRep -> Trace r -> EvalTraceI r a -> Maybe (a, Double)

newtype EvalPredTraceI r a

Constructors

EvalPredTraceI (ReaderT (r HyperRep) (StateT (Trace r, Double) Maybe) a)

runTracePredLogP :: r HyperRep -> Trace r -> EvalPredTraceI r a -> Maybe (a, (Trace r, Double))

evalTracePredLogP :: r HyperRep -> Trace r -> EvalPredTraceI r a -> Maybe (a, Double)

newtype UpdatePriorsI r a

Constructors

UpdatePriorsI (StateT (Trace r, r HyperRep) Maybe a)

getPosterior :: r HyperRep -> Trace r -> UpdatePriorsI r a -> Maybe (r HyperRep)

newtype ShowTraceI r a

Constructors

ShowTraceI (MaybeT (WriterT String (State (Trace r))) a)

showTraceItem :: forall l r. (Show (Support l), Typeable l, Typeable (Support l)) => String -> ShowTraceI r (Support l)

showTrace :: Trace r -> ShowTraceI r a -> (Maybe a, String)

printTrace :: Trace r -> ShowTraceI r a -> IO ()

newtype TraceTraceI r a

Constructors

TraceTraceI (State (Trace r) a)

traceTraceItem :: forall l r. (Show (Support l), Typeable l, Typeable (Support l)) => String -> TraceTraceI r (Support l)

traceTrace :: Trace r -> TraceTraceI r a -> a

data Beta

Constructors

Beta

data Bernoulli

Constructors

Bernoulli

newtype Binomial

Constructors

Binomial Int

data Categorical n

Constructors

Categorical

data Dirichlet n

Constructors

Dirichlet

data Geometric0

Constructors

Geometric0

data Geometric1

Constructors

Geometric1